feat(fw): DDoS hardening, per-source block state machine, paginated blocks + CI release tags
Security review fixes:
- Self-upgrade: SHA-256 verification against the GitHub-reported digest,
  optional minisign signature (fail-closed), secure root-only /run staging
  path (no /tmp TOCTOU); gate POST /upgrade behind api.allow-remote-upgrade.
- Dashboard: vendor preact/htm/hooks locally + strict CSP (no third-party CDN
  on the token-entry page).
- Warn on group/other-readable config; canonicalize CIDRs on API ingest;
  DefaultBodyLimit + ruleset length caps.
- SYN cookie: 64-bit CSPRNG-seeded secret with splitmix64 finalizer; fix
  SYN-proxy accounting undercount before the tail-call.

Source-block redesign:
- Per-source NORMAL/DROPPING state machine with rate hysteresis: a source is
  blocked only while actually over-rate and released on cooldown, not held for
  a blind TTL.
- Aggregate only under trie-space pressure ("merge when running out of
  space"), /32-first, never past a /24 (/48) safety cap (fixes runaway /8
  blocks that blackholed CDN space).
- /blocks paginated (offset/limit/q), sorted with manual blocks pinned to the
  top then auto by pps desc; new dropping/cooling state surfaced in the API +
  dashboard. Refresh docs/dashboard.png (1080p).

CI:
- build.yml: build + push versioned Docker images on release tags (v*),
  keeping :latest on master; assert root Cargo.toml matches the tag.
- lnvps_fw-deb.yml: assert lnvps_fw/Cargo.toml matches the tag before building.

+++ lnvps_fw/lnvps_ebpf/src/maps.rs
@@ -248,9 +248,9 @@ pub static VERIFIED_V4: LruHashMap<[u8; 4], u64> = LruHashMap::with_max_entries(
 #[map]
 pub static VERIFIED_V6: LruHashMap<[u8; 16], u64> = LruHashMap::with_max_entries(MAX_SRC_IPS, 0);
 
-/// Rotating SYN-cookie secret: slot 0 current, slot 1 previous.
+/// Rotating SYN-cookie secret: slot 0 current, slot 1 previous. 64-bit keys.
 #[map]
-pub static COOKIE_SECRET: Array<u32> = Array::with_max_entries(2, 0);
+pub static COOKIE_SECRET: Array<u64> = Array::with_max_entries(2, 0);
 
 /// Tail-call jump table for the SYN-proxy sub-programs (slot 0 = IPv4, slot 1 =
 /// IPv6).
@@ -259,7 +259,7 @@ pub static SYN_PROXY_JUMP: ProgramArray = ProgramArray::with_max_entries(4, 0);
 
 /// Current and previous SYN-cookie secrets (0 if unset).
 #[inline(always)]
-pub fn cookie_secrets() -> (u32, u32) {
+pub fn cookie_secrets() -> (u64, u64) {
     let cur = COOKIE_SECRET
         .get(COOKIE_SECRET_CURRENT)
         .copied()
